test(api): isolate migration ordering contract
